Build Sisco by default (#1481)
* Enable Sisco by default

Enable Sisco. This change requires that we build `libdeflate` ourselves for our Docker base images, because the version in the Alma Linux repository is too old.

* Install libdeflate

We need to install `libdeflate-dev` when building our Docker images

* Install libdeflate for MacOS

* Fix incorrect argument type

The type of the first argument passed to `WriteChunk` is different between `.h` and `.cc` file. This has been fixed.

* Add missing include

Added missing include of header for `std::thread`.

+++ tables/AlternateMans/SiscoWriter.cc
@@ -269,7 +269,7 @@ void SiscoWriter::WriteLoop()
   assert(scheduled_tasks.empty());
 }
 
-void SiscoWriter::WriteChunk(uint64_t uncompressed_size, std::span<const std::byte> data) {
+void SiscoWriter::WriteChunk(size_t uncompressed_size, std::span<const std::byte> data) {
   file_.write(reinterpret_cast<const char*>(&uncompressed_size), sizeof(uncompressed_size));
   const uint64_t compressed_size = data.size();
   file_.write(reinterpret_cast<const char*>(&compressed_size), sizeof(compressed_size));
